Avoid taking 100% CPU.
The fd is set non blocking so the thread can periodically check for the
stop_token. Strictly speaking, the fd could be set to blocking because the
app sends an exit request to the server on termination. This leads to the
server exiting, which leads to the read to return.
However, if the server has an issue and doesn't respect the exit request,
then with a blocking fd, the app would get stuck here.