Use blocking pipes to avoid sleeping between checks
Sleeping incurs delays in the path between the data being sent by the server
and the data being displayed on the screen. Plus, polling increases
resource usages. Instead, makes everything blocking and at the end, for
the case where the server does not respect the EXIT_NOW request, kill it.
When the server dies (either by itself or from being killed) the read
operation from the background thread will stop and thus the program
won't hang